+++ rust/lolfm/src/main.rs
@@ -12,32 +12,27 @@ mod helpers {
 mod models {
   pub mod app_config;
   pub mod cmd;
+  pub mod cmus_event;
   pub mod cmus_status;
   pub mod er;
   pub mod lolfm_event;
-  pub mod played_song;
-  pub mod played_songs;
-  pub mod raw_cmus_event;
   pub mod song;
 }
 
 mod repos {
   pub mod cmd_from_shell;
+  pub mod cmus_events_from_db;
+  pub mod cmus_events_to_db;
   pub mod cmus_tags_from_shell;
   pub mod db_connection;
   pub mod db_init;
   pub mod lolfm_events_to_db;
-  pub mod played_songs_to_db;
-  pub mod raw_cmus_events_delete_from_db;
-  pub mod raw_cmus_events_from_db;
-  pub mod raw_cmus_events_to_db;
   pub mod system_time;
 }
 
 mod transformers {
-  pub mod cmus_tags_to_raw_cmus_events;
-  pub mod raw_cmus_events_to_lolfm_events;
-  pub mod raw_cmus_events_to_played_songs;
+  pub mod cmus_events_to_lolfm_events;
+  pub mod cmus_tags_to_cmus_events;
 }
 
 use actions::get_app_config::get_app_config;
