+++ rust/lolfm/src/models/lolfm_event.rs
@@ -1,6 +1,6 @@
 use crate::models::song::Song;
 
-#[derive(Debug)]
+#[derive(Debug, PartialEq)]
 pub enum LolfmEvent {
   DeleteBefore(i64),
   RecordPlay(i64, Song),
